Fix permuting single-input or single-output recipes to go through all of the permutations.
Because the default recipe is logically the "largest" recipe in I and
R, buildRegistry doesn't see it for computing maxI and maxR; if both
inputs and outputs have permutations then buildRegistry correctly
computes maxI and maxR anyways, since it sees maxI from every R
permutation and so on.

There was a special case hack for input/output-size 2 recipes, we just
needed to extend it to all recipes of this type.